Refactor Test1.java to change variable type from long to int; add new classes for Loan, HomeLoan, CarLoan, Stack, ComplexNum, and TwoDArray with respective functionalities.

+++ Test1.java
@@ -7,7 +7,7 @@ class Test1 {
         x = y = z = 1;
         System.out.println(x + " " + y + " " + z);
 
-        long a = 1;
+        int a = 1;
         switch (a) {
             case 1:
                 System.out.println("1");
